Move duplicated rendering resources into RenderingResourcesSystem
