+++ Features/Sprites/Systems/SpriteRenderSystem.cs
@@ -1,6 +1,4 @@
-using System;
 using System.Numerics;
-using System.Runtime.CompilerServices;
 using Arch.System;
 using Arch.System.SourceGenerator;
 using Diligent;
@@ -11,133 +9,31 @@ using Exanite.GravitationalTetris.Features.Cameras.Components;
 using Exanite.GravitationalTetris.Features.Resources;
 using Exanite.GravitationalTetris.Features.Sprites.Components;
 using Exanite.GravitationalTetris.Features.Transforms.Components;
-using Exanite.ResourceManagement;
 using ValueType = Diligent.ValueType;
 
 namespace Exanite.GravitationalTetris.Features.Sprites.Systems;
 
-public partial class SpriteRenderSystem : EcsSystem, IInitializeSystem, IRenderSystem, IDisposable
+public partial class SpriteRenderSystem : EcsSystem, IInitializeSystem, IRenderSystem
 {
-    private Mesh mesh = null!;
-    private IBuffer uniformBuffer = null!;
-    private IPipelineState pipeline = null!;
-    private IShaderResourceBinding shaderResourceBinding = null!;
-
     private readonly RendererContext rendererContext;
-    private readonly ResourceManager resourceManager;
     private readonly ClearRenderTargetRenderSystem clearRenderTargetRenderSystem;
+    private readonly RenderingResourcesSystem renderingResourcesSystem;
 
-    public SpriteRenderSystem(RendererContext rendererContext, ResourceManager resourceManager, ClearRenderTargetRenderSystem clearRenderTargetRenderSystem)
+    public SpriteRenderSystem(RendererContext rendererContext, ClearRenderTargetRenderSystem clearRenderTargetRenderSystem, RenderingResourcesSystem renderingResourcesSystem)
     {
         this.rendererContext = rendererContext;
-        this.resourceManager = resourceManager;
         this.clearRenderTargetRenderSystem = clearRenderTargetRenderSystem;
+        this.renderingResourcesSystem = renderingResourcesSystem;
     }
 
     public void Initialize()
     {
         clearRenderTargetRenderSystem.ClearColor = Vector4.Zero;
-
-        var renderDevice = rendererContext.RenderDevice;
-        var swapChain = rendererContext.SwapChain;
-
-        mesh = Mesh.Create<VertexPositionUv>("Square mesh", rendererContext, new VertexPositionUv[]
-        {
-            new(new Vector3(-0.5f, -0.5f, 0), new Vector2(0, 0)),
-            new(new Vector3(0.5f, -0.5f, 0), new Vector2(1, 0)),
-            new(new Vector3(0.5f, 0.5f, 0), new Vector2(1, 1)),
-            new(new Vector3(-0.5f, 0.5f, 0), new Vector2(0, 1)),
-        }, new uint[]
-        {
-            2, 1, 0,
-            3, 2, 0,
-        });
-
-        uniformBuffer = rendererContext.RenderDevice.CreateBuffer(new BufferDesc
-        {
-            Name = "Uniform buffer",
-            Size = (ulong)Unsafe.SizeOf<Matrix4x4>(),
-            Usage = Usage.Dynamic,
-            BindFlags = BindFlags.UniformBuffer,
-            CPUAccessFlags = CpuAccessFlags.Write,
-        });
-
-        var vShader = resourceManager.GetResource(BaseMod.SpriteVShader);
-        var pShader = resourceManager.GetResource(BaseMod.SpritePShader);
-
-        pipeline = renderDevice.CreateGraphicsPipelineState(new GraphicsPipelineStateCreateInfo
-        {
-            PSODesc = new PipelineStateDesc
-            {
-                Name = "Sprite PSO",
-                ResourceLayout = new PipelineResourceLayoutDesc
-                {
-                    DefaultVariableType = ShaderResourceVariableType.Static,
-                    Variables = new ShaderResourceVariableDesc[]
-                    {
-                        new()
-                        {
-                            ShaderStages = ShaderType.Pixel,
-                            Name = "Texture",
-                            Type = ShaderResourceVariableType.Mutable,
-                        },
-                    },
-                    ImmutableSamplers = new ImmutableSamplerDesc[]
-                    {
-                        new()
-                        {
-                            Desc = new SamplerDesc
-                            {
-                                MinFilter = FilterType.Point, MagFilter = FilterType.Point, MipFilter = FilterType.Point,
-                                AddressU = TextureAddressMode.Clamp, AddressV = TextureAddressMode.Clamp, AddressW = TextureAddressMode.Clamp,
-                            },
-                            SamplerOrTextureName = "Texture",
-                            ShaderStages = ShaderType.Pixel,
-                        },
-                    },
-                },
-            },
-            Vs = vShader.Value.Handle,
-            Ps = pShader.Value.Handle,
-            GraphicsPipeline = new GraphicsPipelineDesc
-            {
-                InputLayout = VertexPositionUv.Layout,
-                PrimitiveTopology = PrimitiveTopology.TriangleList,
-                RasterizerDesc = new RasterizerStateDesc { CullMode = CullMode.Front },
-                DepthStencilDesc = new DepthStencilStateDesc { DepthEnable = true },
-                BlendDesc = new BlendStateDesc
-                {
-                    RenderTargets = new RenderTargetBlendDesc[]
-                    {
-                        new()
-                        {
-                            BlendEnable = true,
-                            SrcBlend = BlendFactor.SrcAlpha,
-                            DestBlend = BlendFactor.InvSrcAlpha,
-                        },
-                    },
-                },
-                NumRenderTargets = 1,
-                RTVFormats = new[] { swapChain.GetDesc().ColorBufferFormat },
-                DSVFormat = swapChain.GetDesc().DepthBufferFormat,
-            },
-        });
-        pipeline.GetStaticVariableByName(ShaderType.Vertex, "Constants").Set(uniformBuffer, SetShaderResourceFlags.None);
-
-        shaderResourceBinding = pipeline.CreateShaderResourceBinding(true);
     }
 
     public void Render()
     {
         DrawQuery(World);
-    }
-
-    public void Dispose()
-    {
-        shaderResourceBinding.Dispose();
-        pipeline.Dispose();
-        uniformBuffer.Dispose();
-        mesh.Dispose();
     }
 
     [Query]
@@ -151,6 +47,10 @@ public partial class SpriteRenderSystem : EcsSystem, IInitializeSystem, IRenderS
     private void DrawSprites([Data] ref CameraProjectionComponent cameraProjection, ref SpriteComponent sprite, ref TransformComponent transform)
     {
         var deviceContext = rendererContext.DeviceContext;
+        var shaderResourceBinding = renderingResourcesSystem.ShaderResourceBinding;
+        var uniformBuffer = renderingResourcesSystem.UniformBuffer;
+        var pipeline = renderingResourcesSystem.Pipeline;
+        var mesh = renderingResourcesSystem.Mesh;
 
         var texture = sprite.Texture.Value;
         shaderResourceBinding.GetVariableByName(ShaderType.Pixel, "Texture").Set(texture.View, SetShaderResourceFlags.AllowOverwrite);
